requirements.txt

Django>=4.2,<6
pandas>=2.0
openpyxl>=3.1
django-tables2>=2.6
django-filter>=23.0
